+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### ✅ Added
 - Config for bottom placement of reactions
 
+### 🐞 Fixed
+- Video playing after being dismissed on iOS 17.1
+
 # [4.40.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.40.0)
 _October 26, 2023_