+++ CHANGELOG.md
@@ -3,6 +3,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### 🔄 Changed
+
+# [4.41.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.41.0)
+_November 03, 2023_
+
 ### ✅ Added
 - Config for bottom placement of reactions